Improve the app appearance and behavior

+++ app/src/main/java/com/lucazanrosso/randomdraws/ui/EditGroupScreen.kt
@@ -1,10 +1,14 @@
 package com.lucazanrosso.randomdraws.ui
 
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.rounded.Add
@@ -24,7 +28,9 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.lucazanrosso.randomdraws.NavigationDestination
@@ -57,11 +63,14 @@ fun EditGroupScreen(
                     }
                 },
                 actions = {
-                    IconButton(onClick = {
-                        viewModel.upsertItems()
-                        viewModel.deleteItemsFromDb()
-                        navigateBack()
-                    }){
+                    IconButton(
+                        onClick = {
+                            viewModel.upsertItems()
+                            viewModel.deleteItemsFromDb()
+                            navigateBack()
+                        },
+                        enabled = viewModel.newGroupName.isNotEmpty()
+                    ){
                         Icon(
                             imageVector = Icons.Rounded.Done,
                             contentDescription = "Localized description"
@@ -75,14 +84,15 @@ fun EditGroupScreen(
             modifier = modifier
                 .fillMaxWidth()
                 .padding(innerPadding)
-                .padding(start = 16.dp, end = 16.dp)
+                .padding(start = 12.dp, end = 12.dp)
         ) {
             item {
                 OutlinedTextField(
                     value = viewModel.newGroupName,
                     label = { Text(text = "Group name") },
                     onValueChange = { viewModel.newGroupName = it},
-                    modifier = Modifier.fillParentMaxWidth()
+                    keyboardOptions = KeyboardOptions(capitalization = KeyboardCapitalization.Sentences),
+                    modifier = Modifier.fillParentMaxWidth().padding(4.dp)
                 )
             }
 
@@ -90,12 +100,13 @@ fun EditGroupScreen(
                 items = viewModel.itemUiState,
                 key = { _, listItem -> listItem.index })
             { index, item ->
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = modifier.fillMaxWidth()
-                ){
+                Box(
+                    modifier = modifier.fillMaxWidth().padding(4.dp)
+                ) {
+
                     var text by rememberSaveable { mutableStateOf(item.name) }
                     val label = index + 1
+                    var showDelete by rememberSaveable { mutableStateOf(false) }
 
                     OutlinedTextField(
                         value = text,
@@ -104,24 +115,47 @@ fun EditGroupScreen(
                             text = it
                             viewModel.updateItem(index, text)
                         },
-                        modifier = Modifier.weight(1f)
+                        keyboardOptions = KeyboardOptions(capitalization = KeyboardCapitalization.Sentences),
+                        modifier = Modifier
+                            .onFocusChanged { showDelete = it.isFocused }
+                            .fillMaxWidth()
                     )
 
-                    IconButton(
-                        onClick = { viewModel.deleteItem(item) },
-                        modifier = Modifier.padding(top = 8.dp)
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                        modifier = modifier.fillMaxWidth()
                     ) {
-                        Icon(Icons.Rounded.Clear, contentDescription = stringResource(R.string.drag_and_drop))
+                        Spacer(modifier = Modifier.weight(1f))
+
+                        if (showDelete) {
+                            IconButton(
+                                onClick = { viewModel.deleteItem(item) },
+                                modifier = Modifier.padding(top = 11.dp)
+                            ) {
+                                Icon(
+                                    Icons.Rounded.Clear,
+                                    contentDescription = stringResource(R.string.delete_item)
+                                )
+                            }
+                        }
                     }
                 }
             }
 
             item {
-                IconButton(
-                    onClick = { viewModel.addItemToList(viewModel.itemUiState.size) },
-                    modifier = Modifier.padding(top = 8.dp)
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.Center,
+                    modifier = modifier.fillMaxWidth()
                 ) {
-                    Icon(Icons.Rounded.Add, contentDescription = stringResource(R.string.drag_and_drop))
+                    IconButton(
+                        onClick = { viewModel.addItemToList(viewModel.itemUiState.size) },
+                        modifier = Modifier.padding(top = 8.dp)
+                    ) {
+                        Icon(
+                            Icons.Rounded.Add, contentDescription = stringResource(R.string.add_item)
+                        )
+                    }
                 }
             }
         }
